requirements.txt
# Airflow providers
apache-airflow-providers-amazon

# Data processing
polars
pandas
pyarrow
fsspec
s3fs
duckdb

# MinIO client
minio

# Database
psycopg2-binary

# Dashboard
streamlit
plotly
pydeck